requirements.txt
Flask==3.0.3
google-generativeai==0.8.3
openai==1.51.0
requests==2.32.3
beautifulsoup4==4.12.3
scrapy==2.11.0
legacy-cgi
python-dotenv==1.0.0


